renamed folder from jupyterlab-marble-extension to jupyterlab_marble_extension
