--- jupyterlab-marble-extension/src/index.ts
@@ -1,32 +1,0 @@
-import {
-  JupyterFrontEnd,
-  JupyterFrontEndPlugin
-} from '@jupyterlab/application';
-
-import { ISettingRegistry } from '@jupyterlab/settingregistry';
-
-/**
- * Initialization data for the jupyterlab-marble-extension extension.
- */
-const plugin: JupyterFrontEndPlugin<void> = {
-  id: 'jupyterlab-marble-extension:plugin',
-  description: 'A JupyterLab extension to get the session from a Marble node.',
-  autoStart: true,
-  optional: [ISettingRegistry],
-  activate: (app: JupyterFrontEnd, settingRegistry: ISettingRegistry | null) => {
-    console.log('JupyterLab extension jupyterlab-marble-extension is activated!');
-
-    if (settingRegistry) {
-      settingRegistry
-        .load(plugin.id)
-        .then(settings => {
-          console.log('jupyterlab-marble-extension settings loaded:', settings.composite);
-        })
-        .catch(reason => {
-          console.error('Failed to load settings for jupyterlab-marble-extension.', reason);
-        });
-    }
-  }
-};
-
-export default plugin;
