+++ jupyterlab_marble_extension/LICENSE
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2025, DACCS
+Copyright (c) 2025, daccs
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
