+++ jupyterlab_marble_extension/README.md
@@ -1,10 +1,10 @@
 # jupyterlab_marble_extension
 
-[![Github Actions Status](https://github.com/DACCS-Climate/marble-jupyterlab-login-extension/tree/marble-extension/workflows/Build/badge.svg)](https://github.com/DACCS-Climate/marble-jupyterlab-login-extension/tree/marble-extension/actions/workflows/build.yml)
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/DACCS-Climate/marble-jupyterlab-login-extension/tree/marble-extension/main?urlpath=lab)
+[![Github Actions Status](https://github.com/DACCS-Climate/marble-jupyterlab-login-extension/workflows/Build/badge.svg)](https://github.com/DACCS-Climate/marble-jupyterlab-login-extension/actions/workflows/build.yml)
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/DACCS-Climate/marble-jupyterlab-login-extension//main?urlpath=lab)
 
 
-A JupyterLab extension to get the session from a Marble node.
+A JupyterLab extension that gets the session of the marble node
 
 ## Requirements
